+++ src/js/app.js
@@ -14,8 +14,7 @@ import '../css/main.css';
 import { ButtonCountryPTBR } from './components/ButtonPTBR';
 import { ResetButton } from './components/ResetButton';
 import { ButtonCountryEN } from './components/ButtonEN';
-import { loadButtons } from './appBR';
-import { checkResetPTBR } from './checkBR';
+import { loadButtons, checkResetPTBR } from './appBR';
 
 let resetIdiom = '';
