+++ src/js/appBR.js
@@ -1,5 +1,4 @@
 import { clickReset } from './components/ResetButton';
-import { checkIsPTBR } from './checkBR';
 
 export function loadButtons(resetIdiom) {
   clickReset(resetIdiom);
@@ -57,4 +56,23 @@ function mouseHoverEN(e) {
   e.addEventListener('mouseout', () => {
     document.getElementById('button-text-id-en').classList.remove('button-text-show');
   });
+}
+
+export function checkResetPTBR(param) {
+  let resetIdiom = param;
+  if (localStorage.getItem('flagBR') === 'true') {
+    resetIdiom = 'reset-ptbr';
+  } else {
+    resetIdiom = 'reset-english';
+  }
+  return resetIdiom;
+}
+
+export function checkIsPTBR() {
+  if (localStorage.getItem('flagBR') === 'true') {
+    document.querySelector('.root').classList.add('rootbr');
+    document.getElementById('button-ptbr').classList.add('button-spam');
+  } else {
+    document.getElementById('button-en').classList.add('button-spam');
+  }
 }
